Testes/teste_spi/requirements-release.txt
gpiod
spidev
Pillow